fix(login): move App to src/components
